refactor: centralize document type grouping logic
Extract document type grouping logic into a reusable utility function. Refactored related code in pagination, search, and URL endpoints to use this new utility, improving maintainability and reducing duplication.
