+++ fbr/urls.py
@@ -14,6 +14,7 @@ import app.core.views as core_views
 import app.search.views as search_views
 
 from app.cache.manage_cache import rebuild_cache
+from app.search.utils.documents import document_type_groups
 from app.search.utils.search import get_publisher_names, search
 
 urls_logger = logging.getLogger(__name__)
@@ -54,62 +55,7 @@ class DocumentTypesViewSet(viewsets.ViewSet):
     @action(detail=False, methods=["get"], url_path="document-types")
     def document_types(self, request, *args, **kwargs):
         try:
-            import re
-
-            from django.db.models import F
-            from django.db.models.functions import Trim
-
-            from app.search.models import DataResponseModel
-
-            def format_document_type(doc_type):
-                """
-                Format document type string:
-                1. For camel case strings (like EuropeanUnionDecision),
-                   insert spaces between words
-                2. For strings with spaces, capitalize each word
-                """
-                # If already contains spaces, just capitalize each word
-                if " " in doc_type:
-                    return " ".join(
-                        word.capitalize() for word in doc_type.split()
-                    )
-
-                # For camel case, insert spaces before capital letters and
-                # capitalize first letter
-                formatted = re.sub(r"(?<!^)(?=[A-Z])", " ", doc_type)
-                return formatted
-
-            # Get all distinct document types from the DataResponseModel
-            document_types = (
-                DataResponseModel.objects.values(doc_type=Trim(F("type")))
-                .filter(type__isnull=False)
-                .exclude(type__exact="")
-                .distinct()
-                .order_by("doc_type")
-            )
-
-            # Categorize document types
-            non_legislation = []
-            legislation = []
-
-            for item in document_types:
-                if item and item.get("doc_type") is not None:
-                    doc_type = item["doc_type"]
-                    formatted_type = format_document_type(doc_type)
-                    display_item = {
-                        "label": formatted_type,
-                        "name": doc_type,
-                    }
-
-                    if (
-                        "standard" in doc_type.lower()
-                        or "guidance" in doc_type.lower()
-                    ):
-                        # Add to non-legislation
-                        non_legislation.append(display_item)
-                    else:
-                        # Add to legislation
-                        legislation.append(display_item)
+            legislation, non_legislation = document_type_groups()
 
             # Create the response payload
             response_data = {
